backend-server/requirements.txt
Flask
flask-cors
transformers
torch
yt-dlp
openai-whisper